Add Client.StateEvent method with test and example. Move examples to different file
