+++ client_test.go
@@ -8,31 +8,6 @@ import (
 	"testing"
 )
 
-func ExampleClient_BuildURLWithQuery() {
-	cli, _ := NewClient("https://matrix.org", "@example:matrix.org", "abcdef123456")
-	out := cli.BuildURLWithQuery([]string{"sync"}, map[string]string{
-		"filter_id": "5",
-	})
-	fmt.Println(out)
-	// Output: https://matrix.org/_matrix/client/r0/sync?access_token=abcdef123456&filter_id=5
-}
-
-func ExampleClient_BuildURL() {
-	userID := "@example:matrix.org"
-	cli, _ := NewClient("https://matrix.org", userID, "abcdef123456")
-	out := cli.BuildURL("user", userID, "filter")
-	fmt.Println(out)
-	// Output: https://matrix.org/_matrix/client/r0/user/@example:matrix.org/filter?access_token=abcdef123456
-}
-
-func ExampleClient_BuildBaseURL() {
-	userID := "@example:matrix.org"
-	cli, _ := NewClient("https://matrix.org", userID, "abcdef123456")
-	out := cli.BuildBaseURL("_matrix", "client", "r0", "directory", "room", "#matrix:matrix.org")
-	fmt.Println(out)
-	// Output: https://matrix.org/_matrix/client/r0/directory/room/%23matrix:matrix.org?access_token=abcdef123456
-}
-
 func TestClient_LeaveRoom(t *testing.T) {
 	cli := mockClient(func(req *http.Request) (*http.Response, error) {
 		if req.Method == "POST" && req.URL.Path == "/_matrix/client/r0/rooms/!foo:bar/leave" {
@@ -46,6 +21,32 @@ func TestClient_LeaveRoom(t *testing.T) {
 
 	if _, err := cli.LeaveRoom("!foo:bar"); err != nil {
 		t.Fatalf("LeaveRoom: error, got %s", err.Error())
+	}
+}
+
+func TestClient_StateEvent(t *testing.T) {
+	cli := mockClient(func(req *http.Request) (*http.Response, error) {
+		if req.Method == "GET" && req.URL.Path == "/_matrix/client/r0/rooms/!foo:bar/state/m.room.name" {
+			return &http.Response{
+				StatusCode: 200,
+				Body:       ioutil.NopCloser(bytes.NewBufferString(`{"name":"Room Name Goes Here"}`)),
+			}, nil
+		}
+		return nil, fmt.Errorf("unhandled URL: %s", req.URL.Path)
+	})
+
+	content := struct {
+		Name string `json:"name"`
+	}{}
+	expectContent := struct {
+		Name string `json:"name"`
+	}{"Room Name Goes Here"}
+
+	if err := cli.StateEvent("!foo:bar", "m.room.name", "", &content); err != nil {
+		t.Fatalf("StateEvent: error, got %s", err.Error())
+	}
+	if content.Name != expectContent.Name {
+		t.Fatalf("StateEvent: got %s, want %s", content.Name, expectContent.Name)
 	}
 }
